chore(gui): aligned multitenancy related tests in more appropriate file
